Show auto-opened exchangeables as item sources
This introduces a new internal function `recur_replace`, which recursively expands a particular drop table.

The expansion only happens on exchangeables that are automatically opened.

The purpose of this recursive expansion is to capture more broad item sources.

For example, before this change, the Crossbow was only shown to drop from glitch and weaponbox. After this change, Crossbows are now shown to be obtainable from glitch, weaponbox, and lostearring1

Notably, this function *does not* show lostearring1 as a source of weaponbox, as per a previous comment from mid-December. This behavior was intended not to exist